# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduMaster 360 - School Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marks:
    get:
      operationId: listMarks
      tags: [marks]
      summary: List marks
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: term
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Marks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mark"
    post:
      operationId: createMark
      tags: [marks]
      summary: Record a mark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mark"
  /marks/{id}:
    patch:
      operationId: updateMark
      tags: [marks]
      summary: Update a mark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mark"
    delete:
      operationId: deleteMark
      tags: [marks]
      summary: Delete a mark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ATTENDANCE ───────────────────────────────────────────────────────
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: term
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Create fee record / record payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/{id}:
    patch:
      operationId: updateFeeRecord
      tags: [fees]
      summary: Update fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
...
  /rankings/students:
    get:
      operationId: getStudentRankings
      tags: [rankings]
      summary: Get student rankings (class or school-wide)
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Student rankings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRanking"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Mark:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        subject:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: ["number", "null"]
        grade:
          type: ["string", "null"]
        term:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, subject, score, maxScore, term, createdAt]
    MarkInput:
      type: object
      properties:
        studentId:
          type: integer
        subject:
          type: string
        score:
          type: number
        maxScore:
          type: number
        term:
          type: string
        teacherId:
          type: ["integer", "null"]
        remarks:
          type: ["string", "null"]
      required: [studentId, subject, score, maxScore, term]
    MarkUpdate:
      type: object
      properties:
        score:
          type: number
        maxScore:
          type: number
        remarks:
          type: ["string", "null"]

    # ─── ATTENDANCE ───────────────────────────────────────────────────
...
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        term:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
          enum: [paid, partial, unpaid]
        bursarId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, term, totalAmount, amountPaid, balance, status, createdAt]
    FeeInput:
      type: object
      properties:
        studentId:
          type: integer
        term:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        notes:
          type: ["string", "null"]
      required: [studentId, term, totalAmount, amountPaid]
    FeeUpdate:
      type: object
      properties:
        amountPaid:
          type: number
        notes:
          type: ["string", "null"]
...
    StudentRanking:
      type: object
      properties:
        studentId:
          type: integer
        name:
          type: string
        className:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        overallScore:
          type: number
        academicScore:
          type: number
        disciplineScore:
          type: number
        attendanceScore:
          type: number
        rank:
          type: integer
        classRank:
          type: ["integer", "null"]
        trend:
          type: string
          enum: [up, down, stable]
      required: [studentId, name, overallScore, academicScore, disciplineScore, attendanceScore, rank, trend]
...